+++ rust-crates/sns-sdk/src/non_blocking/mod.rs
@@ -5,4 +5,3 @@ pub mod record_v1;
 pub mod record_v2;
 pub mod resolve;
 pub mod subdomain;
-pub(crate) mod tld;